requirements.txt
# Core async HTTP and web scraping - OPTIMIZED VERSIONS
aiohttp[speedups]>=3.9.0  # Use speedups version for better performance
beautifulsoup4>=4.12.0
lxml>=4.9.0  # Faster XML/HTML parsing

# Data processing - PERFORMANCE OPTIMIZED
pandas>=2.0.0  # Newer version with performance improvements
numpy>=1.24.0

# Google Cloud BigQuery - WITH PERFORMANCE PACKAGE
google-cloud-bigquery>=3.11.0
pandas-gbq>=0.19.0  # This will eliminate the warning and improve performance

# Timezone handling
pytz>=2023.3

# CSV handling optimization
chardet>=5.2.0

# Performance monitoring and profiling (optional)
psutil>=5.9.0  # For memory monitoring
cProfile  # Built-in, for performance profiling

# Optional: for better process titles (Unix-like systems)
setproctitle>=1.3.0; sys_platform != "win32"

# Windows-specific packages
pywin32>=306; sys_platform == "win32"

# Optional: for even better performance (advanced users)
# uvloop>=0.17.0; sys_platform != "win32"  # Faster event loop for Linux
# aiodns>=3.0.0  # Faster DNS resolution